requirements.txt
adafruit-circuitpython-dht
PyMySQL
python-dotenv
lgpio